// ClientApp/src/store/messagebox/messagebox.types.ts
export enum MESSAGEBOX_ACTION_TYPES {
    SET_IS_ARTIFICIALINTELLIGENCE_OPEN = 'messagebox/SET_IS_ARTIFICIALINTELLIGENCE_OPEN',
    SET_IS_DEVICES_OPEN = 'messagebox/SET_IS_DEVICES_OPEN',
    SET_IS_POST_OPEN = 'messagebox/SET_IS_POST_OPEN',
    SET_IS_SOLARSYSTEM_OPEN = 'messagebox/SET_IS_SOLARSYSTEM_OPEN',
    SET_IS_EDITOR_OPEN = 'messagebox/SET_IS_EDITOR_OPEN',
    SET_IS_BUILDER_OPEN = 'messagebox/SET_IS_BUILDER_OPEN',
    SET_IS_MARAUDERS_OPEN = 'messagebox/SET_IS_MARAUDERS_OPEN',
    SET_IS_MESSAGES_OPEN = 'messagebox/SET_IS_MESSAGES_OPEN'
}